+++ sdk/core/azure-core/src/http/retry_policy.cpp
@@ -158,7 +158,7 @@ Context inline CreateRetryContext(Context const& parent)
   int retryCount = 0;
   if (parent.HasKey(RetryKey))
   {
-    retryCount = parent[RetryKey].Get<int>() + 1;
+    retryCount = parent.Get<int>(RetryKey) + 1;
   }
   return parent.WithValue(RetryKey, retryCount);
 }
@@ -175,7 +175,7 @@ int RetryPolicy::GetRetryNumber(Context const& context)
     // ...
     return -1;
   }
-  return context[RetryKey].Get<int>();
+  return context.Get<int>(RetryKey);
 }
 
 std::unique_ptr<RawResponse> RetryPolicy::Send(