+++ sdk/core/azure-core/test/ut/policy.cpp
@@ -93,7 +93,8 @@ TEST(Policy, throwWhenNoTransportPolicy)
   Azure::Core::Http::_internal::HttpPipeline pipeline(policies);
   Azure::Core::Http::Url url("");
   Azure::Core::Http::Request request(Azure::Core::Http::HttpMethod::Get, url);
-  EXPECT_THROW(pipeline.Send(request, Azure::Core::GetApplicationContext()), std::invalid_argument);
+  EXPECT_THROW(
+      pipeline.Send(request, Azure::Core::Context::GetApplicationContext()), std::invalid_argument);
 }
 
 TEST(Policy, throwWhenNoTransportPolicyMessage)
@@ -115,7 +116,7 @@ TEST(Policy, throwWhenNoTransportPolicyMessage)
 
   try
   {
-    pipeline.Send(request, Azure::Core::GetApplicationContext());
+    pipeline.Send(request, Azure::Core::Context::GetApplicationContext());
   }
   catch (const std::invalid_argument& ex)
   {
@@ -142,7 +143,7 @@ TEST(Policy, ValuePolicy)
 
   Request request(HttpMethod::Get, Url("https:://www.example.com"));
 
-  pipeline.Send(request, GetApplicationContext());
+  pipeline.Send(request, Context::GetApplicationContext());
 
   auto headers = request.GetHeaders();
   auto queryParams = request.GetUrl().GetQueryParameters();
@@ -161,7 +162,7 @@ TEST(Policy, RetryPolicyCounter)
   retryCounterState = 0;
 
   // Check when there's no info about retry on the context
-  auto initialContext = GetApplicationContext();
+  auto initialContext = Context::GetApplicationContext();
   EXPECT_EQ(-1, RetryPolicy::GetRetryNumber(initialContext));
 
   // Pipeline with retry test
@@ -197,5 +198,5 @@ TEST(Policy, RetryPolicyRetryCycle)
 
   HttpPipeline pipeline(policies);
   Request request(HttpMethod::Get, Url("url"));
-  pipeline.Send(request, GetApplicationContext());
+  pipeline.Send(request, Context::GetApplicationContext());
 }