+++ sdk/core/azure-core/test/ut/simplified_header.cpp
@@ -31,7 +31,7 @@ TEST(SimplifiedHeader, core)
   EXPECT_NO_THROW(Azure::Core::CaseInsensitiveMap imap);
   EXPECT_NO_THROW(Azure::Core::CaseInsensitiveSet iset);
   EXPECT_NO_THROW(Azure::Core::Context c);
-  EXPECT_NO_THROW(Azure::Core::DateTime(2020, 11, 03, 15, 30, 44));
+  EXPECT_NO_THROW(Azure::DateTime(2020, 11, 03, 15, 30, 44));
   EXPECT_NO_THROW(Azure::ETag e);
   EXPECT_NO_THROW(Azure::Core::Base64Decode("foo"));
   EXPECT_NO_THROW(Azure::Core::Cryptography::Md5Hash m);
@@ -44,7 +44,7 @@ TEST(SimplifiedHeader, core)
   EXPECT_NO_THROW(Azure::Core::Uuid::CreateUuid());
   EXPECT_NO_THROW(Azure::Core::RequestFailedException("foo"));
   EXPECT_NO_THROW(Azure::Core::OperationStatus("foo"));
-  EXPECT_NO_THROW(Azure::Core::Version::VersionString());
+  EXPECT_NO_THROW(Azure::Core::PackageVersion::VersionString());
 
   {
     std::vector<uint8_t> buffer(10);