+++ sdk/storage/azure-storage-blobs/test/ut/block_blob_client_test.cpp
@@ -114,7 +114,7 @@ namespace Azure { namespace Storage { namespace Test {
   {
     const int64_t downloadLength = 1024;
     Blobs::DownloadBlobOptions options;
-    options.Range = Azure::Core::Http::Range();
+    options.Range = Azure::Core::Http::HttpRange();
     options.Range.GetValue().Offset = 0;
     options.Range.GetValue().Length = downloadLength;
     options.RangeHashAlgorithm = HashAlgorithm::Md5;
@@ -152,7 +152,7 @@ namespace Azure { namespace Storage { namespace Test {
       EXPECT_TRUE(IsValidTime(res->LastAccessedOn.GetValue()));
     }
     {
-      Azure::Core::DateTime lastAccessedOn;
+      Azure::DateTime lastAccessedOn;
 
       Azure::Storage::Blobs::ListBlobsSinglePageOptions options;
       options.Prefix = m_blobName;
@@ -196,7 +196,7 @@ namespace Azure { namespace Storage { namespace Test {
     EXPECT_EQ(res->BlobType, Azure::Storage::Blobs::Models::BlobType::BlockBlob);
 
     Azure::Storage::Blobs::DownloadBlobOptions options;
-    options.Range = Core::Http::Range();
+    options.Range = Core::Http::HttpRange();
     options.Range.GetValue().Offset = 0;
     EXPECT_THROW(blockBlobClient.Download(options), StorageException);
     options.Range.GetValue().Length = 1;
@@ -481,7 +481,7 @@ namespace Azure { namespace Storage { namespace Test {
       options.TransferOptions.Concurrency = concurrency;
       if (offset.HasValue() || length.HasValue())
       {
-        options.Range = Core::Http::Range();
+        options.Range = Core::Http::HttpRange();
         options.Range.GetValue().Offset = offset.GetValue();
         options.Range.GetValue().Length = length;
       }
@@ -553,7 +553,7 @@ namespace Azure { namespace Storage { namespace Test {
       options.TransferOptions.Concurrency = concurrency;
       if (offset.HasValue() || length.HasValue())
       {
-        options.Range = Core::Http::Range();
+        options.Range = Core::Http::HttpRange();
         options.Range.GetValue().Offset = offset.GetValue();
         options.Range.GetValue().Length = length;
       }
@@ -638,7 +638,7 @@ namespace Azure { namespace Storage { namespace Test {
       // buffer not big enough
       Blobs::DownloadBlobToOptions options;
       options.TransferOptions.Concurrency = c;
-      options.Range = Core::Http::Range();
+      options.Range = Core::Http::HttpRange();
       options.Range.GetValue().Offset = 1;
       for (int64_t length : {1ULL, 2ULL, 4_KB, 5_KB, 8_KB, 11_KB, 20_KB})
       {
@@ -789,7 +789,7 @@ namespace Azure { namespace Storage { namespace Test {
       EXPECT_TRUE(ReadFile(tempFilename).empty());
       DeleteFile(tempFilename);
 
-      options.Range = Core::Http::Range();
+      options.Range = Core::Http::HttpRange();
       options.Range.GetValue().Offset = 0;
       EXPECT_THROW(
           blockBlobClient.DownloadTo(emptyContent.data(), static_cast<std::size_t>(8_MB), options),