+++ sdk/storage/azure-storage-common/inc/azure/storage/common/storage_switch_to_secondary_policy.hpp
@@ -13,7 +13,7 @@ namespace Azure { namespace Storage { namespace _detail {
   static constexpr const char* SecondaryHostReplicaStatusKey
       = "AzureSdkStorageSecondaryHostReplicaStatusKey";
 
-  struct SecondaryHostReplicaStatus : public Azure::Core::ValueBase
+  struct SecondaryHostReplicaStatus
   {
     bool replicated = true;
   };