+++ sdk/storage/azure-storage-common/src/storage_per_retry_policy.cpp
@@ -24,8 +24,8 @@ namespace Azure { namespace Storage { namespace _detail {
       // add x-ms-date header in RFC1123 format
       request.SetHeader(
           HttpHeaderXMsDate,
-          Core::DateTime(std::chrono::system_clock::now())
-              .ToString(Azure::Core::DateTime::DateFormat::Rfc1123));
+          DateTime(std::chrono::system_clock::now())
+              .ToString(Azure::DateTime::DateFormat::Rfc1123));
     }
 
     return nextHttpPolicy.Send(request, ctx);